Add estimated wait time, and expected time
Should probably put your time zone (matching the PC running the app) in the string format, so that users can expect their new trade requests to be fulfilled at.

Pair this with the waited.txt and you can see if your queue is growing or shrinking, or if your estimation factor needs updating.