Add ACNH drop bot
It's a dumb bot, but it works fine. Automatically wraps items and drops them at the player's feet.

Using: console app; be in-game with gates open in an area you want visitors to pick up items from your feet.

No, the bot will not travel to other people's islands.
No, the bot will not provide metadata on who is on the island.
No, the bot doesn't prevent people from spamming requests.

The bot requires users to be responsible and only request items when they are ready for pickup -- ie, right in front of the player-bot on their island.