Less dumb GUI log truncation
RichTextBox.Lines' implementation builds a new list of strings every `get`, which chops up the backing `Text` property into each line.

For noisy or long running bots, constantly calling this (and with lots of existing lines) will be a lot of pressure on the GC, slowing the execution environment.

Just truncate if the new string will exceed the length, and lock it in the event multiple threads trample each other.